requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
assemblyai==0.20.0
google-generativeai==0.7.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
alembic==1.13.0
tenacity==8.2.3
